Fixed problem with the note html element having an id of "noteundefined" when first created.
Added some note unit tests.

Fixed problem with the code that disables creating a second click on a
shift-doubleclick disallowing creating any notes after the first click.

Added some better error checking for Ajax.Request 

Made sure that __populate() never overwrites fields that have changed since the
state of the object was last sent to the server.  This was manifesting itself
as a problem where typing into a note right after it was created wasn't getting
saved, because the data returned from insert() contained an empty "contents"
field, which was overwriting the newly set contents field.

Made sure that update gets called properly if insert isn't finished when it's
called.